feat: Helpers - saveNewComment, formatComment, selectAuthor
Create format comment helper to use in the Api request
saveNewComment to post a new comment. Inside saveNewComment
return the object resulted from adding the formatted comment
and the response from the api. This obj is the complete
comment. Create selectAuthor to randomly create an author
username for the comment.