feat: Navi, Sidebar, App - add state in App to control when sidebar is displayed
Add prop in App state and a method to toggle when Sidebar is
displayed or not. Pass that method and state prop down to Sidebar
and Navi. In navi, when user clicks the menu button, it opens the
sidebar. The sidebar in xs and sm devices is displayed only when user
clicks the menu button in the navbar. The sidebar has a dismiss button
to close it.