feat: HomeView - render all posts and posts per category in HomeView
In App render the Homeview component when user routes to
the category posts instead of rendering CategoryView
(which is not used anymore, can be deleted). In Home
View grab the params coming from the route and make
a category prop with that category . If the category
prop is null, that is, all posts are beeing displayed
render all the posts, if the user routes to one
category filter and show only the posts of that
category.